refactor: move fixtures to correct dir
